1. Moved 3D view to it's own component 2. Started to build the PuzzleInfo stuff 3. Cubes can generate for different types
