+++ src/app/app.module.ts
@@ -6,12 +6,14 @@ import { AppComponent } from './app.component';
 import { MapComponent } from './map/map.component';
 import { PuzzleViewComponent } from './puzzle-view/puzzle-view.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { ThreeDViewComponent } from './three-dview/three-dview.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     MapComponent,
-    PuzzleViewComponent
+    PuzzleViewComponent,
+    ThreeDViewComponent
   ],
   imports: [
     BrowserModule,
