+++ src/app/three-dview/three-dview.component.html
@@ -1,0 +1,3 @@
+<div id="canvasWrap" #canvasHolder>
+  <canvas #canvas id="canvas" style="height: 500px; width: 100%"></canvas>
+</div>
