+++ src/app/three-dview/three-dview.component.scss
@@ -1,0 +1,3 @@
+#canvasWrap {
+  border: 1px solid black;
+}
